# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Enhara API
  version: 0.1.0
  description: API do MVP para veículos, telemetria, saúde, viagens, diagnósticos, alertas, simulação e eventos SSE.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Mesmo host da aplicação ou proxy local
security: []
...
paths:
...
  /api/vehicles/{vehicleId}:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      tags: [Vehicles]
      summary: Obtém um veículo
      operationId: getVehicle
      responses:
        "200":
          description: Veículo encontrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Vehicle" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/vehicle-data/fipe/brands:
    get:
      tags: [Vehicle Data]
      summary: Lista marcas reais disponíveis no catálogo FIPE
      operationId: listFipeBrands
      parameters:
        - $ref: "#/components/parameters/FipeVehicleType"
      responses:
        "200":
          description: Marcas retornadas pela BrasilAPI
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FipeOption" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "503": { $ref: "#/components/responses/ProviderUnavailable" }
  /api/vehicle-data/fipe/brands/{brandCode}/models:
    get:
      tags: [Vehicle Data]
      summary: Lista modelos/versões FIPE de uma marca
      operationId: listFipeModels
      parameters:
        - $ref: "#/components/parameters/FipeVehicleType"
        - $ref: "#/components/parameters/FipeBrandCode"
      responses:
        "200":
          description: Modelos retornados pela BrasilAPI
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FipeOption" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "503": { $ref: "#/components/responses/ProviderUnavailable" }
  /api/vehicle-data/fipe/brands/{brandCode}/models/{modelCode}/years:
    get:
      tags: [Vehicle Data]
      summary: Lista combinações de ano e combustível de um modelo FIPE
      operationId: listFipeYears
      parameters:
        - $ref: "#/components/parameters/FipeVehicleType"
        - $ref: "#/components/parameters/FipeBrandCode"
        - $ref: "#/components/parameters/FipeModelCode"
      responses:
        "200":
          description: Anos e combustíveis retornados pela BrasilAPI
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FipeOption" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "503": { $ref: "#/components/responses/ProviderUnavailable" }
...
  /api/vehicles/{vehicleId}/notes:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      tags: [Notes]
      summary: Lista notas e lembretes cadastrados pelo usuário
      operationId: listVehicleNotes
      parameters:
        - name: includeCompleted
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Notas ordenadas pela última alteração
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VehicleNote" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Notes]
      summary: Cadastra uma nota ou lembrete fornecido pelo usuário
      operationId: createVehicleNote
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VehicleNoteInput" }
      responses:
        "201":
          description: Nota criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VehicleNote" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vehicles/{vehicleId}/notes/{noteId}:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
      - $ref: "#/components/parameters/NoteId"
    put:
      tags: [Notes]
      summary: Edita uma nota cadastrada pelo usuário
      operationId: updateVehicleNote
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VehicleNoteInput" }
      responses:
        "200":
          description: Nota atualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VehicleNote" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Notes]
      summary: Exclui uma nota cadastrada pelo usuário
      operationId: deleteVehicleNote
      responses:
        "204": { description: Nota excluída }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vehicles/{vehicleId}/notes/{noteId}/complete:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
      - $ref: "#/components/parameters/NoteId"
    post:
      tags: [Notes]
      summary: Marca uma nota ou lembrete como concluído
      operationId: completeVehicleNote
      responses:
        "200":
          description: Nota concluída
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VehicleNote" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vehicles/{vehicleId}/notes/{noteId}/reopen:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
      - $ref: "#/components/parameters/NoteId"
    post:
      tags: [Notes]
      summary: Reabre uma nota ou lembrete concluído
      operationId: reopenVehicleNote
      responses:
        "200":
          description: Nota reaberta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VehicleNote" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/vehicles/{vehicleId}/simulation:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      tags: [Simulation]
      summary: Consulta estado do simulador auxiliar
      operationId: getSimulationStatus
      responses:
        "200":
          description: Estado atual
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimulationStatus" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vehicles/{vehicleId}/simulation/start:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    post:
      tags: [Simulation]
      summary: Inicia o simulador auxiliar
      operationId: startSimulation
      responses:
        "200":
          description: Simulador iniciado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimulationStatus" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vehicles/{vehicleId}/simulation/obd:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      tags: [Simulation]
      summary: Consulta capabilities e estado atual da ECU/OBD simulada
      description: DTCs deste recurso pertencem à ECU simulada e não são findings nem alertas do Enhara.
      operationId: getSimulatedObdState
      responses:
        "200":
          description: Snapshot da sessão OBD simulada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimulatedObdSnapshot" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vehicles/{vehicleId}/simulation/stop:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    post:
      tags: [Simulation]
      summary: Para o simulador auxiliar
      operationId: stopSimulation
      responses:
        "200":
          description: Simulador parado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimulationStatus" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vehicles/{vehicleId}/simulation/tick:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    post:
      tags: [Simulation]
      summary: Gera uma amostra determinística no cenário selecionado
      operationId: tickSimulation
      responses:
        "200":
          description: Amostra gerada, diagnosticada e persistida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IngestionResult" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vehicles/{vehicleId}/simulation/scenario/{scenario}:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
      - name: scenario
        in: path
        required: true
        schema: { $ref: "#/components/schemas/SimulationScenario" }
    post:
      tags: [Simulation]
      summary: Altera o cenário sem editar código
      operationId: setSimulationScenario
      responses:
        "200":
          description: Cenário selecionado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimulationStatus" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vehicles/{vehicleId}/simulation/profile/{profile}:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
      - name: profile
        in: path
        required: true
        schema: { $ref: "#/components/schemas/SimulationProfile" }
    post:
      tags: [Simulation]
      summary: Reinicia a ECU simulada com outro perfil de capabilities
      operationId: setSimulationProfile
      responses:
        "200":
          description: Perfil selecionado e estado reiniciado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimulationStatus" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  parameters:
    VehicleId:
      name: vehicleId
      in: path
      required: true
      schema: { type: string, format: uuid }
    NoteId:
      name: noteId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    FipeVehicleType:
      name: vehicleType
      in: query
      schema: { $ref: "#/components/schemas/FipeVehicleType", default: CAR }
    FipeBrandCode:
      name: brandCode
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9]+$" }
    FipeModelCode:
      name: modelCode
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9]+$" }
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
    ProviderUnavailable:
      description: Provider externo indisponível; o restante da API permanece funcional
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
    CreateVehicle:
      type: object
      additionalProperties: false
      required:
        [name, manufacturer, model, modelYear, licensePlate, odometerKm]
      properties:
        name: { type: string, maxLength: 80, examples: [Enhara Demo Car] }
        vin:
          type: [string, "null"]
          pattern: "^[A-HJ-NPR-Z0-9]{17}$"
          description: Opcional; pode ser informado manualmente ou observado posteriormente por OBD real.
        manufacturer: { type: string, maxLength: 80 }
        model: { type: string, maxLength: 80 }
        modelYear: { type: integer, minimum: 1886, maximum: 2100 }
        licensePlate: { type: string, pattern: "^[A-Z0-9-]{6,10}$" }
        odometerKm: { type: number, minimum: 0 }
    Vehicle:
      allOf:
        - $ref: "#/components/schemas/CreateVehicle"
        - type: object
          required: [id, createdAt]
          properties:
            id: { type: string, format: uuid }
            createdAt: { type: string, format: date-time }
...
    FipeVehicleType:
      type: string
      enum: [CAR, MOTORCYCLE, TRUCK]
    FipeOption:
      type: object
      required: [code, label]
      properties:
        code: { type: string }
        label: { type: string }
...
    TelemetryInput:
      type: object
      additionalProperties: false
      required:
        [
          speedKph,
          rpm,
          engineTempC,
          engineLoadPercent,
          throttlePositionPercent,
          batteryVoltage,
          fuelLevelPercent,
        ]
      properties:
        recordedAt:
          {
            type: string,
            format: date-time,
            description: Opcional; o backend usa o instante atual quando ausente.,
          }
        speedKph: { type: number, minimum: 0, maximum: 350 }
        rpm: { type: integer, minimum: 0, maximum: 12000 }
        engineTempC: { type: number, minimum: -50, maximum: 180 }
        engineLoadPercent: { type: number, minimum: 0, maximum: 100 }
        throttlePositionPercent: { type: number, minimum: 0, maximum: 100 }
        batteryVoltage: { type: number, minimum: 0, maximum: 30 }
        fuelLevelPercent: { type: number, minimum: 0, maximum: 100 }
        latitude: { type: [number, "null"], minimum: -90, maximum: 90 }
        longitude: { type: [number, "null"], minimum: -180, maximum: 180 }
        source: { type: string, enum: [SIMULATED_OBD, SIMULATOR, MOBILE, API], default: API }
    Telemetry:
      allOf:
        - $ref: "#/components/schemas/TelemetryInput"
        - type: object
          required: [id, vehicleId, recordedAt, source]
          properties:
            id: { type: integer, format: int64 }
            vehicleId: { type: string, format: uuid }
...
    IngestionResult:
      type: object
      required: [telemetry, diagnostics, newAlerts]
      properties:
        telemetry: { $ref: "#/components/schemas/Telemetry" }
        diagnostics:
          type: array
          items: { $ref: "#/components/schemas/Diagnostic" }
        newAlerts:
          type: array
          items: { $ref: "#/components/schemas/Alert" }
...
    Severity:
      type: string
      enum: [INFO, WARNING, CRITICAL]
    Diagnostic:
      type: object
      required: [id, vehicleId, code, description, severity, status, detectedAt]
      properties:
        id: { type: string, format: uuid }
        vehicleId: { type: string, format: uuid }
        telemetryId: { type: [integer, "null"], format: int64 }
        code:
          {
            type: string,
            enum:
              [ENGINE_TEMPERATURE_HIGH, BATTERY_VOLTAGE_LOW, ENGINE_OVERSPEED],
          }
        description: { type: string }
        severity: { $ref: "#/components/schemas/Severity" }
        status: { type: string, enum: [ACTIVE, RESOLVED] }
        detectedAt: { type: string, format: date-time }
        resolvedAt: { type: [string, "null"], format: date-time }
    Alert:
      type: object
      required:
        [id, vehicleId, type, severity, title, message, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        vehicleId: { type: string, format: uuid }
        telemetryId: { type: [integer, "null"], format: int64 }
        type:
          {
            type: string,
            enum: [ENGINE_OVERHEAT, LOW_BATTERY, LOW_FUEL, ENGINE_OVERSPEED],
          }
        severity: { $ref: "#/components/schemas/Severity" }
        title: { type: string }
        message: { type: string }
        status: { type: string, enum: [OPEN, ACKNOWLEDGED] }
        createdAt: { type: string, format: date-time }
        acknowledgedAt: { type: [string, "null"], format: date-time }
...
    VehicleNoteCategory:
      type: string
      enum: [MAINTENANCE, DOCUMENTATION, GENERAL]
    VehicleNoteStatus:
      type: string
      enum: [OPEN, COMPLETED]
    VehicleNoteInput:
      type: object
      additionalProperties: false
      required: [title, description, category]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, minLength: 1, maxLength: 1000 }
        category: { $ref: "#/components/schemas/VehicleNoteCategory" }
        dueAt: { type: [string, "null"], format: date-time }
    VehicleNote:
      allOf:
        - $ref: "#/components/schemas/VehicleNoteInput"
        - type: object
          required: [id, vehicleId, status, overdue, createdAt, updatedAt]
          properties:
            id: { type: string, format: uuid }
            vehicleId: { type: string, format: uuid }
            status: { $ref: "#/components/schemas/VehicleNoteStatus" }
            overdue: { type: boolean }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
            completedAt: { type: [string, "null"], format: date-time }
    SimulationScenario:
      type: string
      enum: [NORMAL, OVERHEAT, LOW_VOLTAGE, MISFIRE, LOW_BATTERY]
      description: LOW_BATTERY é alias legado; clientes novos devem usar LOW_VOLTAGE.
    SimulationProfile:
      type: string
      enum: [COMPACT_GASOLINE, COMPACT_GASOLINE_LIMITED]
    SimulationStatus:
      type: object
      required: [vehicleId, running, scenario, generatedSamples, profile]
      properties:
        vehicleId: { type: string, format: uuid }
        running: { type: boolean }
        scenario: { $ref: "#/components/schemas/SimulationScenario" }
        generatedSamples: { type: integer, format: int64, minimum: 0 }
        profile: { $ref: "#/components/schemas/SimulationProfile" }
    ObdCapabilityStatus:
      type: string
      enum: [SUPPORTED, UNSUPPORTED, UNKNOWN]
    ObdAvailabilityStatus:
      type: string
      enum: [SUPPORTED, SUPPORTED_NO_DATA, UNSUPPORTED, UNKNOWN, STALE]
    ObdCapability:
      type: object
      required: [key, service, pid, status, availability]
      properties:
        key: { type: string }
        service: { type: string, pattern: "^[0-9A-F]{2}$" }
        pid: { type: string, pattern: "^[0-9A-F]{2}$" }
        unit: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/ObdCapabilityStatus" }
        availability: { $ref: "#/components/schemas/ObdAvailabilityStatus" }
    ObdLivePidValue:
      type: object
      required: [key, service, pid, value, unit, availability, observedAt]
      properties:
        key: { type: string }
        service: { type: string, pattern: "^[0-9A-F]{2}$" }
        pid: { type: string, pattern: "^[0-9A-F]{2}$" }
        value: { type: number }
        unit: { type: string }
        availability: { $ref: "#/components/schemas/ObdAvailabilityStatus" }
        observedAt: { type: string, format: date-time }
    SimulatedDriverInput:
      type: object
      required: [throttlePercent, brakePercent]
      properties:
        throttlePercent: { type: number, minimum: 0, maximum: 100 }
        brakePercent: { type: number, minimum: 0, maximum: 100 }
    SimulatedVehicleState:
      type: object
      required:
        [speedKph, rpm, gear, shifting, engineLoadPercent, coolantTemperatureC,
         intakeAirTemperatureC, controlModuleVoltage, fuelLevelPercent]
      properties:
        speedKph: { type: number, minimum: 0 }
        rpm: { type: integer, minimum: 0 }
        gear: { type: integer, minimum: 1 }
        shifting: { type: boolean }
        shiftedFromGear: { type: [integer, "null"], minimum: 1 }
        shiftedToGear: { type: [integer, "null"], minimum: 1 }
        engineLoadPercent: { type: number, minimum: 0, maximum: 100 }
        coolantTemperatureC: { type: number }
        intakeAirTemperatureC: { type: number }
        controlModuleVoltage: { type: number, minimum: 0 }
        fuelLevelPercent: { type: number, minimum: 0, maximum: 100 }
    ObdFreezeFrame:
      type: object
      required: [capturedAt, values]
      properties:
        capturedAt: { type: string, format: date-time }
        values:
          type: array
          items: { $ref: "#/components/schemas/ObdLivePidValue" }
    ObdDtcStatus:
      type: string
      enum: [PENDING, CONFIRMED, PERMANENT]
    SimulatedObdDtc:
      type: object
      required: [code, description, statuses, active, firstDetectedAt, lastDetectedAt]
      properties:
        code: { type: string, pattern: "^[PBCU][0-9A-F]{4}$" }
        description: { type: string }
        statuses:
          type: array
          uniqueItems: true
          items: { $ref: "#/components/schemas/ObdDtcStatus" }
        active: { type: boolean }
        firstDetectedAt: { type: string, format: date-time }
        lastDetectedAt: { type: string, format: date-time }
        freezeFrame:
          oneOf:
            - $ref: "#/components/schemas/ObdFreezeFrame"
            - type: "null"
    ObdReadiness:
      type: object
      required: [monitor, status]
      properties:
        monitor:
          type: string
          enum: [MISFIRE, FUEL_SYSTEM, COMPREHENSIVE_COMPONENT, CATALYST, OXYGEN_SENSOR]
        status:
          type: string
          enum: [READY, NOT_READY, NOT_SUPPORTED]
    SimulatedVehicleInformation:
      type: object
      required: [vin, source]
      properties:
        vin: { type: string, pattern: "^[A-HJ-NPR-Z0-9]{17}$" }
        source: { type: string, const: SIMULATED_OBD }
    SimulatedObdSnapshot:
      type: object
      required:
        [profile, scenario, capabilities, driverInput, vehicleState, liveData, dtcs,
         milOn, readiness, elapsedSeconds]
      properties:
        profile: { $ref: "#/components/schemas/SimulationProfile" }
        scenario: { $ref: "#/components/schemas/SimulationScenario" }
        capabilities:
          type: array
          items: { $ref: "#/components/schemas/ObdCapability" }
        driverInput: { $ref: "#/components/schemas/SimulatedDriverInput" }
        vehicleState: { $ref: "#/components/schemas/SimulatedVehicleState" }
        liveData:
          type: array
          items: { $ref: "#/components/schemas/ObdLivePidValue" }
        dtcs:
          type: array
          description: Códigos mantidos pela ECU simulada; não são findings do Enhara.
          items: { $ref: "#/components/schemas/SimulatedObdDtc" }
        milOn: { type: boolean }
        readiness:
          type: array
          items: { $ref: "#/components/schemas/ObdReadiness" }
        vehicleInformation:
          oneOf:
            - $ref: "#/components/schemas/SimulatedVehicleInformation"
            - type: "null"
        elapsedSeconds: { type: integer, format: int64, minimum: 0 }
    ApiError:
      type: object
      required: [timestamp, status, error, message, path, fieldErrors]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }